+++ server/src/main/java/org/allaymc/server/world/AllayWorld.java
@@ -45,10 +45,12 @@ import java.util.concurrent.atomic.AtomicBoolean;
  */
 @Slf4j
 public class AllayWorld implements World {
+
     // Send the time to client every 12 seconds
     protected static final int TIME_SENDING_INTERVAL = 12 * 20;
     protected static final int MAX_PACKETS_HANDLE_COUNT_AT_ONCE = Server.SETTINGS.networkSettings().maxSyncedPacketsHandleCountAtOnce();
     protected static final boolean ENABLE_INDEPENDENT_NETWORK_THREAD = Server.SETTINGS.networkSettings().enableIndependentNetworkThread();
+    protected static final boolean TICK_DIMENSION_IN_PARALLEL = Server.SETTINGS.worldSettings().tickDimensionInParallel();
 
     @Getter
     protected final String name;
@@ -177,10 +179,19 @@ public class AllayWorld implements World {
         if (!ENABLE_INDEPENDENT_NETWORK_THREAD) {
             handleSyncPackets(null);
         }
+
         tickTime(currentTick);
         tickWeather();
         scheduler.tick();
-        getDimensions().values().forEach(d -> ((AllayDimension) d).tick(currentTick));
+
+        if (TICK_DIMENSION_IN_PARALLEL) {
+            dimensionMap.values().parallelStream().forEach(d -> ((AllayDimension) d).tick(currentTick));
+        } else {
+            for (var dimension : dimensionMap.values()) {
+                ((AllayDimension) dimension).tick(currentTick);
+            }
+        }
+
         worldStorage.tick(currentTick);
     }